extracted mod data to new benchmarking/crypto/mock/tests files
